requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
Pygments==2.19.2
PyMuPDF==1.27.1
pynndescent==0.6.0
pyparsing==3.3.2
pypdf==6.7.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.19
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
umap-learn==0.5.11
